refactor: move period parsing helpers to time_period.vectorized
The extract_month/extract_week/extract_year/season_column helpers are not
threshold-specific -- the outbreak-detection metrics use them too -- so they
belong in the time_period package rather than under assessment/thresholds.
Not re-exported from the package __init__ to keep CLI startup imports lean.
