+++ chap_core/time_period/vectorized.py
